+++ ams/core/service.py
@@ -8,7 +8,7 @@ from typing import Callable, Type
 import numpy as np
 import scipy.sparse as spr
 
-from andes.core.service import BaseService, BackRef, RefFlatten  # NOQA
+from andes.core.service import BaseService
 
 from ams.opt.omodel import Param
 
@@ -268,14 +268,14 @@ class NumOp(ROperationService):
     def __init__(self,
                  u: Callable,
                  fun: Callable,
-                 args: dict = dict(),
+                 args: dict = None,
                  name: str = None,
                  tex_name: str = None,
                  unit: str = None,
                  info: str = None,
                  vtype: Type = None,
                  rfun: Callable = None,
-                 rargs: dict = dict(),
+                 rargs: dict = None,
                  expand_dims: int = None,
                  array_out=True,
                  no_parse: bool = False,
@@ -285,9 +285,9 @@ class NumOp(ROperationService):
                          info=info, vtype=vtype, u=u,
                          no_parse=no_parse, sparse=sparse)
         self.fun = fun
-        self.args = args
+        self.args = {} if args is None else args
         self.rfun = rfun
-        self.rargs = rargs
+        self.rargs = {} if rargs is None else rargs
         self.expand_dims = expand_dims
         self.array_out = array_out
 
@@ -347,7 +347,7 @@ class NumExpandDim(NumOp):
     def __init__(self,
                  u: Callable,
                  axis: int = 0,
-                 args: dict = {},
+                 args: dict = None,
                  name: str = None,
                  tex_name: str = None,
                  unit: str = None,
@@ -414,14 +414,14 @@ class NumOpDual(NumOp):
                  u: Callable,
                  u2: Callable,
                  fun: Callable,
-                 args: dict = {},
+                 args: dict = None,
                  name: str = None,
                  tex_name: str = None,
                  unit: str = None,
                  info: str = None,
                  vtype: Type = None,
                  rfun: Callable = None,
-                 rargs: dict = {},
+                 rargs: dict = None,
                  expand_dims: int = None,
                  array_out=True,
                  no_parse: bool = False,
@@ -545,14 +545,14 @@ class NumHstack(NumOp):
     def __init__(self,
                  u: Callable,
                  ref: Callable,
-                 args: dict = {},
+                 args: dict = None,
                  name: str = None,
                  tex_name: str = None,
                  unit: str = None,
                  info: str = None,
                  vtype: Type = None,
                  rfun: Callable = None,
-                 rargs: dict = {},
+                 rargs: dict = None,
                  no_parse: bool = False,
                  sparse: bool = False,):
         super().__init__(name=name, tex_name=tex_name, unit=unit,
@@ -638,7 +638,7 @@ class ZonalSum(NumOp):
                  info: str = None,
                  vtype: Type = None,
                  rfun: Callable = None,
-                 rargs: dict = {},
+                 rargs: dict = None,
                  no_parse: bool = False,
                  sparse: bool = False,):
         super().__init__(name=name, tex_name=tex_name, unit=unit,
@@ -722,7 +722,7 @@ class VarSelect(NumOp):
                  info: str = None,
                  vtype: Type = None,
                  rfun: Callable = None,
-                 rargs: dict = {},
+                 rargs: dict = None,
                  array_out: bool = True,
                  no_parse: bool = False,
                  sparse: bool = False,
@@ -815,7 +815,7 @@ class VarReduction(NumOp):
                  info: str = None,
                  vtype: Type = None,
                  rfun: Callable = None,
-                 rargs: dict = {},
+                 rargs: dict = None,
                  no_parse: bool = False,
                  sparse: bool = False,
                  **kwargs):
@@ -872,7 +872,7 @@ class RampSub(NumOp):
                  info: str = None,
                  vtype: Type = None,
                  rfun: Callable = None,
-                 rargs: dict = {},
+                 rargs: dict = None,
                  no_parse: bool = False,
                  sparse: bool = False,):
         super().__init__(name=name, tex_name=tex_name, unit=unit,