+++ ams/opt/omodel.py
@@ -6,6 +6,7 @@ import logging
 from typing import Any, Optional, Union
 from collections import OrderedDict
 import re
+import ast
 
 import numpy as np
 import scipy.sparse as spr
@@ -95,6 +96,67 @@ class OptzBase:
             return None
 
 
+class ExpressionCalc(OptzBase):
+    """
+    Expression for calculation.
+    """
+
+    def __init__(self,
+                 name: Optional[str] = None,
+                 info: Optional[str] = None,
+                 unit: Optional[str] = None,
+                 var: Optional[str] = None,
+                 e_str: Optional[str] = None,
+                 ):
+        OptzBase.__init__(self, name=name, info=info, unit=unit)
+        self.optz = None
+        self.var = var
+        self.e_str = e_str
+        self.code = None
+
+    def parse(self, no_code=True):
+        """
+        Parse the Expression.
+
+        Parameters
+        ----------
+        no_code : bool, optional
+            Flag indicating if the code should be shown, True by default.
+        """
+        # parse the expression str
+        sub_map = self.om.rtn.syms.sub_map
+        code_expr = self.e_str
+        for pattern, replacement in sub_map.items():
+            try:
+                code_expr = re.sub(pattern, replacement, code_expr)
+            except TypeError as e:
+                logger.error(f"Error in parsing expr <{self.name}>.")
+                raise e
+        # store the parsed expression str code
+        self.code = code_expr
+        code_expr = "self.optz = " + code_expr
+        msg = f"Parse ExpressionCalc <{self.name}>: {self.e_str} "
+        logger.debug(msg)
+        if not no_code:
+            logger.info(f"<{self.name}> code: {code_expr}")
+        # execute the expression
+        exec(code_expr, globals(), locals())
+        return True
+
+    @property
+    def v(self):
+        """
+        Return the CVXPY expression value.
+        """
+        if self.optz is None:
+            return None
+        else:
+            return self.optz.value
+
+    def __repr__(self):
+        return f'{self.__class__.__name__}: {self.name}'
+
+
 class Param(OptzBase):
     """
     Base class for parameters used in a routine.
@@ -107,8 +169,8 @@ class Param(OptzBase):
         True to set the parameter as non-negative.
     nonpos: bool, optional
         True to set the parameter as non-positive.
-    complex: bool, optional
-        True to set the parameter as complex.
+    cplx: bool, optional
+        True to set the parameter as complex, avoiding the use of `complex`.
     imag: bool, optional
         True to set the parameter as imaginary.
     symmetric: bool, optional
@@ -136,7 +198,7 @@ class Param(OptzBase):
                  no_parse: Optional[bool] = False,
                  nonneg: Optional[bool] = False,
                  nonpos: Optional[bool] = False,
-                 complex: Optional[bool] = False,
+                 cplx: Optional[bool] = False,
                  imag: Optional[bool] = False,
                  symmetric: Optional[bool] = False,
                  diag: Optional[bool] = False,
@@ -155,7 +217,7 @@ class Param(OptzBase):
 
         self.config.add(OrderedDict((('nonneg', nonneg),
                                      ('nonpos', nonpos),
-                                     ('complex', complex),
+                                     ('complex', cplx),
                                      ('imag', imag),
                                      ('symmetric', symmetric),
                                      ('diag', diag),
@@ -238,7 +300,7 @@ class Var(OptzBase):
         Non-negative variable
     nonpos : bool, optional
         Non-positive variable
-    complex : bool, optional
+    cplx : bool, optional
         Complex variable
     imag : bool, optional
         Imaginary variable
@@ -283,7 +345,7 @@ class Var(OptzBase):
                  horizon=None,
                  nonneg: Optional[bool] = False,
                  nonpos: Optional[bool] = False,
-                 complex: Optional[bool] = False,
+                 cplx: Optional[bool] = False,
                  imag: Optional[bool] = False,
                  symmetric: Optional[bool] = False,
                  diag: Optional[bool] = False,
@@ -319,7 +381,7 @@ class Var(OptzBase):
 
         self.config.add(OrderedDict((('nonneg', nonneg),
                                      ('nonpos', nonpos),
-                                     ('complex', complex),
+                                     ('complex', cplx),
                                      ('imag', imag),
                                      ('symmetric', symmetric),
                                      ('diag', diag),
@@ -425,9 +487,9 @@ class Constraint(OptzBase):
         A mathematical expression representing the constraint.
     info : str, optional
         Additional informational text about the constraint.
-    type : str, optional
-        The type of constraint, which determines the mathematical relationship.
-        Possible values include 'uq' (inequality, default) and 'eq' (equality).
+    is_eq : str, optional
+        Flag indicating if the constraint is an equality constraint. False indicates
+        an inequality constraint in the form of `<= 0`.
 
     Attributes
     ----------
@@ -441,11 +503,11 @@ class Constraint(OptzBase):
                  name: Optional[str] = None,
                  e_str: Optional[str] = None,
                  info: Optional[str] = None,
-                 type: Optional[str] = 'uq',
+                 is_eq: Optional[str] = False,
                  ):
         OptzBase.__init__(self, name=name, info=info)
         self.e_str = e_str
-        self.type = type
+        self.is_eq = is_eq
         self.is_disabled = False
         self.dual = None
         self.code = None
@@ -475,11 +537,9 @@ class Constraint(OptzBase):
         self.code = code_constr
         # parse the constraint type
         code_constr = "self.optz=" + code_constr
-        if self.type not in ['uq', 'eq']:
-            raise ValueError(f'Constraint type {self.type} is not supported.')
-        code_constr += " <= 0" if self.type == 'uq' else " == 0"
+        code_constr += " == 0" if self.is_eq else " <= 0"
         msg = f"Parse Constr <{self.name}>: {self.e_str} "
-        msg += " <= 0" if self.type == 'uq' else " == 0"
+        msg += " == 0" if self.is_eq else " <= 0"
         logger.debug(msg)
         if not no_code:
             logger.info(f"<{self.name}> code: {code_constr}")
@@ -515,7 +575,7 @@ class Constraint(OptzBase):
 
         try:
             logger.debug(f"Value code: {code}")
-            out = eval(code)
+            out = ast.literal_eval(code)
             return out
         except Exception as e:
             logger.error(f"Error in calculating constr <{self.name}>.")
@@ -603,7 +663,7 @@ class Objective(OptzBase):
 
         try:
             logger.debug(f"Value code: {code}")
-            out = eval(code)
+            out = ast.literal_eval(code)
             return out
         except Exception as e:
             logger.error(f"Error in calculating obj <{self.name}>.")
@@ -805,6 +865,8 @@ class OModel:
                 msg += ", ".join(constrs_skip)
             logger.debug(msg)
             exec(code_prob, globals(), locals())
+            for key, val in self.rtn.exprs.items():
+                val.parse()
 
         _, s_setup = elapsed(t_setup)
         self.initialized = True
@@ -819,7 +881,7 @@ class OModel:
         """
         return self.__class__.__name__
 
-    def __register_attribute(self, key, value):
+    def _register_attribute(self, key, value):
         """
         Register a pair of attributes to OModel instance.
 
@@ -834,7 +896,7 @@ class OModel:
             self.params[key] = value
 
     def __setattr__(self, __name: str, __value: Any):
-        self.__register_attribute(__name, __value)
+        self._register_attribute(__name, __value)
         super().__setattr__(__name, __value)
 
     def update(self, params):